Add files via upload
This is a project that I did in December based on the game of Blackjack. I practiced performing data collection, cleaning, preparing, clustering analysis, logistic regression, and I created a confusion matrix using R Studio. In this, I figured out that it is beneficial to count cards, even with a slight amount of human error. The dataset was created by myself, with over 400 hands of blackjack played with basic rules like H17, 6:5, no surrender, no DAS, and 2 decks. Most of these rules do not benefit the player, except for having 2 decks. I would go on, but it is all in the project file if your would like to learn more!